fix(trusty-review): fix silent mid-file hunk truncation in render_for_prompt
The inner `break` in the hunk loop of `FilteredDiff::render_for_prompt`
only exited the inner hunk loop, not the outer file loop.  This allowed
subsequent files to be rendered after a half-rendered file whose hunks
overflowed the char budget — with no marker indicating the drop.

The reviewer LLM would see File A (partial: N of M hunks), then File B,
File C … in full, with no signal that File A was incomplete.  This is a
form of silent diff truncation distinct from the already-fixed raw-diff
cap (MAX_DIFF_CHARS, #624).

Fix:
- Use a labeled loop `'files:` and `break 'files` from the inner hunk
  budget check so the outer loop is exited when the budget is hit.
- Track `budget_exceeded: bool`; when set, append a loud
  `[RENDER TRUNCATED — char budget (N) reached …]` marker before the
  noise summary so the LLM cannot miss that content was omitted.
- Reserve `TRUNC_MARKER_RESERVE = 120` chars in the budget check so the
  marker itself never causes a large overshoot.
- Split tests into a sibling `models_tests.rs` (models.rs was 613 lines
  after the new tests; now 360+278, both under the 500-line cap).

New tests:
  render_for_prompt_mid_file_hunk_overflow_loud_not_silent — regression
    guard for the inner-break bug.
  render_for_prompt_no_truncation_marker_when_fits — no false-positive.
  render_for_prompt_marker_does_not_cause_large_overflow — overhead bound.

Closes: #622 (the render_for_prompt silent-truncation vector).
Related: #624 (MAX_DIFF_CHARS cap raise + DiffAnalyzer noise filter).
